+++ pact-jvm-matchers/src/test/groovy/au/com/dius/pact/matchers/HeaderMatcherSpec.groovy
@@ -1,6 +1,6 @@
 package au.com.dius.pact.matchers
 
-import au.com.dius.pact.model.matchingrules.MatchingRules
+import au.com.dius.pact.model.matchingrules.MatchingRulesImpl
 import au.com.dius.pact.model.matchingrules.RegexMatcher
 import scala.None$
 import scala.collection.JavaConversions
@@ -11,24 +11,24 @@ class HeaderMatcherSpec extends Specification {
   def "matching headers - be true when headers are equal"() {
     expect:
     HeaderMatcher.compareHeader('HEADER', 'HEADER', 'HEADER',
-      new MatchingRules()) == None$.MODULE$
+      new MatchingRulesImpl()) == None$.MODULE$
   }
 
   def "matching headers - be false when headers are not equal"() {
     expect:
     HeaderMatcher.compareHeader('HEADER', 'HEADER', 'HEADSER',
-      new MatchingRules()) != None$.MODULE$
+      new MatchingRulesImpl()) != None$.MODULE$
   }
 
   def "matching headers - exclude whitespace from the comparison"() {
     expect:
     HeaderMatcher.compareHeader('HEADER', 'HEADER1, HEADER2,   3', 'HEADER1,HEADER2,3',
-      new MatchingRules()) == None$.MODULE$
+      new MatchingRulesImpl()) == None$.MODULE$
   }
 
   def "matching headers - delegate to a matcher when one is defined"() {
     given:
-    def matchers = new MatchingRules()
+    def matchers = new MatchingRulesImpl()
     matchers.addCategory('header').addRule('HEADER', new RegexMatcher('.*'))
 
     expect:
@@ -38,36 +38,36 @@ class HeaderMatcherSpec extends Specification {
   def "matching headers - content type header - be true when headers are equal"() {
     expect:
     HeaderMatcher.compareHeader('CONTENT-TYPE', 'application/json;charset=UTF-8',
-      'application/json; charset=UTF-8', new MatchingRules()) == None$.MODULE$
+      'application/json; charset=UTF-8', new MatchingRulesImpl()) == None$.MODULE$
   }
 
   def "matching headers - content type header - be false when headers are not equal"() {
     expect:
     HeaderMatcher.compareHeader('CONTENT-TYPE', 'application/json;charset=UTF-8',
-      'application/pdf;charset=UTF-8', new MatchingRules()) != None$.MODULE$
+      'application/pdf;charset=UTF-8', new MatchingRulesImpl()) != None$.MODULE$
   }
 
   def "matching headers - content type header - be false when charsets are not equal"() {
     expect:
     HeaderMatcher.compareHeader('CONTENT-TYPE', 'application/json;charset=UTF-8',
-      'application/json;charset=UTF-16', new MatchingRules()) != None$.MODULE$
+      'application/json;charset=UTF-16', new MatchingRulesImpl()) != None$.MODULE$
   }
 
   def "matching headers - content type header - be false when other parameters are not equal"() {
     expect:
     HeaderMatcher.compareHeader('CONTENT-TYPE', 'application/json;declaration="<950118.AEB0@XIson.com>"',
-      'application/json;charset=UTF-8', new MatchingRules()) != None$.MODULE$
+      'application/json;charset=UTF-8', new MatchingRulesImpl()) != None$.MODULE$
   }
 
   def "matching headers - content type header - be true when the charset is missing from the expected header"() {
     expect:
     HeaderMatcher.compareHeader('CONTENT-TYPE', 'application/json',
-      'application/json ; charset=UTF-8', new MatchingRules()) == None$.MODULE$
+      'application/json ; charset=UTF-8', new MatchingRulesImpl()) == None$.MODULE$
   }
 
   def "matching headers - content type header - delegate to any defined matcher"() {
     given:
-    def matchers = new MatchingRules()
+    def matchers = new MatchingRulesImpl()
     matchers.addCategory('header').addRule('CONTENT-TYPE', new RegexMatcher('[a-z]+\\/[a-z]+'))
 
     expect: