+++ pact-jvm-matchers/src/test/groovy/au/com/dius/pact/matchers/TypeMatcherSpec.groovy
@@ -2,7 +2,7 @@ package au.com.dius.pact.matchers
 
 import au.com.dius.pact.model.OptionalBody
 import au.com.dius.pact.model.Request
-import au.com.dius.pact.model.matchingrules.MatchingRules
+import au.com.dius.pact.model.matchingrules.MatchingRulesImpl
 import au.com.dius.pact.model.matchingrules.NumberTypeMatcher
 import spock.lang.Specification
 
@@ -12,7 +12,7 @@ class TypeMatcherSpec extends Specification {
 
   def 'match integers should accept integer values'() {
     given:
-    def matchingRules = new MatchingRules()
+    def matchingRules = new MatchingRulesImpl()
     matchingRules.addCategory('body').addRule('$.value', new NumberTypeMatcher(NumberTypeMatcher.NumberType.INTEGER))
     def expected = new Request('get', '/', null, null, OptionalBody.body('{"value": 123}'), matchingRules)
     def actual = new Request('get', '/', null, null, OptionalBody.body('{"value": 456}'), null)
@@ -26,7 +26,7 @@ class TypeMatcherSpec extends Specification {
 
   def 'match integers should not match null values'() {
     given:
-    def matchingRules = new MatchingRules()
+    def matchingRules = new MatchingRulesImpl()
     matchingRules.addCategory('body').addRule('$.value', new NumberTypeMatcher(NumberTypeMatcher.NumberType.INTEGER))
     def expected = new Request('get', '/', null, null, OptionalBody.body('{"value": 123}'), matchingRules)
     def actual = new Request('get', '/', null, null, OptionalBody.body('{"value": null}'), null)
@@ -40,7 +40,7 @@ class TypeMatcherSpec extends Specification {
 
   def 'match integers should fail for non-integer values'() {
     given:
-    def matchingRules = new MatchingRules()
+    def matchingRules = new MatchingRulesImpl()
     matchingRules.addCategory('body').addRule('$.value', new NumberTypeMatcher(NumberTypeMatcher.NumberType.INTEGER))
     def expected = new Request('get', '/', null, null, OptionalBody.body('{"value": 123}'), matchingRules)
     def actual = new Request('get', '/', null, null, OptionalBody.body('{"value": 123.10}'), null)
@@ -54,7 +54,7 @@ class TypeMatcherSpec extends Specification {
 
   def 'match decimal should accept decimal values'() {
     given:
-    def matchingRules = new MatchingRules()
+    def matchingRules = new MatchingRulesImpl()
     matchingRules.addCategory('body').addRule('$.value', new NumberTypeMatcher(NumberTypeMatcher.NumberType.DECIMAL))
     def expected = new Request('get', '/', null, null, OptionalBody.body('{"value": 123.10}'), matchingRules)
     def actual = new Request('get', '/', null, null, OptionalBody.body('{"value": 456.20}'), null)
@@ -68,7 +68,7 @@ class TypeMatcherSpec extends Specification {
 
   def 'match decimal should handle null values'() {
     given:
-    def matchingRules = new MatchingRules()
+    def matchingRules = new MatchingRulesImpl()
     matchingRules.addCategory('body').addRule('$.value', new NumberTypeMatcher(NumberTypeMatcher.NumberType.DECIMAL))
     def expected = new Request('get', '/', null, null, OptionalBody.body('{"value": 123.10}'), matchingRules)
     def actual = new Request('get', '/', null, null, OptionalBody.body('{"value": null}'), null)
@@ -82,7 +82,7 @@ class TypeMatcherSpec extends Specification {
 
   def 'match decimal should fail for non-decimal values'() {
     given:
-    def matchingRules = new MatchingRules()
+    def matchingRules = new MatchingRulesImpl()
     matchingRules.addCategory('body').addRule('$.value', new NumberTypeMatcher(NumberTypeMatcher.NumberType.DECIMAL))
     def expected = new Request('get', '/', null, null, OptionalBody.body('{"value": 123.10}'), matchingRules)
     def actual = new Request('get', '/', null, null, OptionalBody.body('{"value": 123}'), null)