+++ pact-jvm-model/src/main/groovy/au/com/dius/pact/model/Request.groovy
@@ -4,6 +4,7 @@ import au.com.dius.pact.model.generators.Category
 import au.com.dius.pact.model.generators.Generator
 import au.com.dius.pact.model.generators.Generators
 import au.com.dius.pact.model.matchingrules.MatchingRules
+import au.com.dius.pact.model.matchingrules.MatchingRulesImpl
 import groovy.transform.Canonical
 import org.jetbrains.annotations.NotNull
 
@@ -11,7 +12,7 @@ import org.jetbrains.annotations.NotNull
  * Request made by a consumer to a provider
  */
 @Canonical
-class Request extends HttpPart implements Comparable {
+class Request extends BaseRequest implements Comparable {
   private static final String COOKIE_KEY = 'cookie'
 
   public static final String DEFAULT_METHOD = 'GET'
@@ -22,7 +23,7 @@ class Request extends HttpPart implements Comparable {
   Map<String, List<String>> query = [:]
   Map<String, String> headers = [:]
   OptionalBody body = OptionalBody.missing()
-  MatchingRules matchingRules = new MatchingRules()
+  MatchingRules matchingRules = new MatchingRulesImpl()
   Generators generators = new Generators()
 
   static Request fromMap(Map map) {
@@ -32,7 +33,7 @@ class Request extends HttpPart implements Comparable {
       query = map.query ?: [:]
       headers = map.headers ?: [:]
       body = map.containsKey('body') ? OptionalBody.body(map.body) : OptionalBody.missing()
-      matchingRules = MatchingRules.fromMap(map.matchingRules)
+      matchingRules = MatchingRulesImpl.fromMap(map.matchingRules)
       generators = Generators.fromMap(map.generators)
       it
     }