+++ pact-jvm-provider-junit/src/main/java/au/com/dius/pact/provider/junit/target/Target.java
@@ -2,6 +2,9 @@ package au.com.dius.pact.provider.junit.target;
 
 import au.com.dius.pact.model.Interaction;
 import au.com.dius.pact.model.PactSource;
+import au.com.dius.pact.provider.ProviderVerifier;
+
+import java.util.function.BiConsumer;
 
 /**
  * Run {@link Interaction} and perform response verification
@@ -18,4 +21,9 @@ public interface Target {
      * @param source Source of the Pact interaction
      */
     void testInteraction(String consumer, Interaction interaction, PactSource source);
+
+    /**
+     * Add a callback to receive the test interaction result
+     */
+    void addResultCallback(BiConsumer<Boolean, ProviderVerifier> callback);
 }