+++ pact-jvm-provider/src/main/kotlin/au/com/dius/pact/provider/ProviderVerifier.kt
@@ -4,6 +4,7 @@ import au.com.dius.pact.model.BrokerUrlSource
 import au.com.dius.pact.model.Pact
 import au.com.dius.pact.provider.broker.PactBrokerClient
 import au.com.dius.pact.provider.broker.com.github.kittinunf.result.Result
+import groovy.lang.GroovyObjectSupport
 import mu.KotlinLogging
 import java.util.function.Function
 
@@ -24,10 +25,10 @@ fun reportVerificationResults(pact: Pact, result: Boolean, version: String, clie
   }
 }
 
-open class ProviderVerifierBase {
+open class ProviderVerifierBase : GroovyObjectSupport() {
 
-  var projectHasProperty = Function<String, Boolean> { false }
-  var projectGetProperty = Function<String, String?> { null }
+  var projectHasProperty = Function<String, Boolean> { name -> !System.getProperty(name).isNullOrEmpty() }
+  var projectGetProperty = Function<String, String?> { name -> System.getProperty(name) }
 
   /**
    * This will return true if the pact.verifier.publishResults property is present and has the value of "false"